Migrate endpoints to rest api specification

+++ jams/util/helper.py
@@ -15,29 +15,68 @@ def prep_delete_event_location(event_location_id):
     event_location = EventLocation.query.filter_by(id=event_location_id).first()
     event_sessions = event_location.sessions
     for event_session in event_sessions:
-        prep_delete_session(event_session.id)
+        if not prep_delete_session(event_session.id):
+            return False
         db.session.delete(event_session)
     
     # TODO: Remove anything else here in future (Not needed at the time of writing)
     db.session.commit()
+
+    # Check if there are no sessions for this event_location
+    if len(event_location.sessions) > 0:
+        # There are still sessions, so return false
+        return False
+    
+    # Everything is removed, so return true
+    return True
+
 
 def prep_delete_event_Timeslot(event_timeslot_id):
     event_timeslot = EventTimeslot.query.filter_by(id=event_timeslot_id).first()
     event_sessions = event_timeslot.sessions
     for event_session in event_sessions:
-        prep_delete_session(event_session.id)
+        if not prep_delete_session(event_session.id):
+            return False
         db.session.delete(event_session)
     
     # TODO: Remove anything else here in future (Not needed at the time of writing)
     db.session.commit()
 
+    # Check if there are no sessions for this event_timeslot
+    if len(event_timeslot.sessions) > 0:
+        # There are still sessions, so return false
+        return False
+    
+    # Everything is removed, so return true
+    return True
+
 def prep_delete_session(session_id):
+    ############ This is not needed anymore, but will be needed in the future. So will leave it #######
     session = Session.query.filter_by(id=session_id).first()
-    session_workshop = session.session_workshop
-    if session_workshop is not None:
-        db.session.delete(session_workshop)
-        # TODO: Remove anything else here in future (Not needed at the time of writing)
-        db.session.commit()
+    
+    # TODO: Remove anything else here in future (Not needed at the time of writing)
+    
+    # Everything is removed, so return true
+    return True
+
+def prep_delete_role(role):
+    # Get all the users for a specified role
+    users = role.users
+
+    # Iterate through each user and remove the role from them
+    for user in users:
+        user.remove_roles(role.id)
+    
+    # Commit the changes to the DB
+    db.session.commit()
+
+    # Check if no users have the role
+    if role.users is not None:
+        # Users still have the role, so return false
+        return False
+    
+    # Everything is removed, so return true
+    return True
     
 
 def reorder_ids(id_list, target_id, new_index):
